requirements.txt
opencv-python
torch
ultralytics
pyserial
geopy
